Implemented new cache system retrieval for: - Configuration _ CommandsConfiguration - GUIsConfiguration Also added a new automatic update for these configurations to simplify the data conversion in PTM updates.

+++ src/main/java/me/thegabro/playtimemanager/Updates/Version34to341Updater.java
@@ -113,67 +113,18 @@ public class Version34to341Updater {
     }
 
     private void recreateConfigFile() {
-        File configFile = new File(plugin.getDataFolder(), "config.yml");
-        FileConfiguration config = YamlConfiguration.loadConfiguration(configFile);
-        String permissionsManagerPlugin = config.getString("permissions-manager-plugin");
-        String datetimeFormat = config.getString("datetime-format");
-        String prefix = config.getString("prefix");
-        String playtimeSelfMessage = config.getString("playtime-self-message");
-        String playtimeOthersMessage = config.getString("playtime-others-message");
-        long goalsCheckRate = config.getLong("goal-check-rate");
-        boolean goalsCheckVerbose = config.getBoolean("goal-check-verbose");
-        String streakResetSchedule = config.getString("streak-reset-schedule");
-        String streakResetTimezone = config.getString("reset-schedule-timezone");
-        boolean rewardsCheckScheduleActivation = config.getBoolean("rewards-check-schedule-activation");
-        boolean streakCheckVerbose = config.getBoolean("streak-check-verbose");
-        boolean resetJoinStreakEnabled = config.getBoolean("reset-joinstreak.enabled");
-        int resetJoinstreakMissedJoins = config.getInt("reset-joinstreak.missed-joins");
-        String joinWarnClaimMessage = config.getString("join-warn-claim-message");
-        String joinWarnAutoclaimMessage = config.getString("join-warn-autoclaim-message");
-        String joinUnclaimedPreviousMessage = config.getString("join-unclaimed-previous-message");
-        boolean placeHoldersErrors = config.getBoolean("placeholders.enable-errors");
-        String placeHoldersDefaultMSG = config.getString("placeholders.default-message");
-
-        configFile.delete();
-
-        Configuration newConfig = new Configuration(
-                plugin.getDataFolder(),
-                "config",
-                true,
-                true
-        );
-
-        newConfig.setPermissionsManagerPlugin(permissionsManagerPlugin);
-        newConfig.setDateTimeFormat(datetimeFormat);
-        newConfig.setPluginChatPrefix(prefix);
-        newConfig.setPlaytimeSelfMessage(playtimeSelfMessage);
-        newConfig.setPlaytimeOthersMessage(playtimeOthersMessage);
-        newConfig.setGoalsCheckRate(goalsCheckRate);
-        newConfig.setGoalsCheckVerbose(goalsCheckVerbose);
-        newConfig.setStreakResetSchedule(streakResetSchedule);
-        newConfig.setStreakTimeZone(streakResetTimezone);
-        newConfig.setRewardsCheckScheduleActivation(rewardsCheckScheduleActivation);
-        newConfig.setStreakCheckVerbose(streakCheckVerbose);
-        newConfig.setJoinStreakResetActivation(resetJoinStreakEnabled);
-        newConfig.setJoinStreakResetMissesAllowed(resetJoinstreakMissedJoins);
-        newConfig.setJoinClaimMessage(joinWarnClaimMessage);
-        newConfig.setJoinAutoClaimMessage(joinWarnAutoclaimMessage);
-        newConfig.setJoinCantClaimMessage(joinUnclaimedPreviousMessage);
-        newConfig.setPlaceholdersEnableErrors(placeHoldersErrors);
-        newConfig.setPlaceholdersDefaultMessage(placeHoldersDefaultMSG);
-
-        newConfig.reload();
-
-        plugin.setConfiguration(newConfig);
+        Configuration.getInstance().updateConfig(true);
     }
 
     private void updateCommandsConfigFile() {
-        File commandsConfigFile = new File(plugin.getDataFolder(), "Customizations/Commands/commands-config.yml");
+        File commandsConfigFile = new File(plugin.getDataFolder(), "Translations/Commands/commands-config.yml");
 
-        // If the file doesn't exist, skip the update
-        if (!commandsConfigFile.exists()) {
-            plugin.getLogger().info("Commands configuration file not found. Skipping commands config update.");
-            return;
+        File parentDir = commandsConfigFile.getParentFile();
+        if (!parentDir.exists()) {
+            if (!parentDir.mkdirs()) {
+                plugin.getLogger().severe("Failed to create directory structure for commands configuration file");
+                return;
+            }
         }
 
         try {
@@ -216,9 +167,11 @@ public class Version34to341Updater {
 
 
             // Delete the original file
-            if (!commandsConfigFile.delete()) {
-                plugin.getLogger().warning("Could not delete old commands configuration file");
-                return;
+            if (commandsConfigFile.exists()) {
+                if (!commandsConfigFile.delete()) {
+                    plugin.getLogger().warning("Could not delete old commands configuration file");
+                    return;
+                }
             }
 
             // Create new configuration with updated structure
